+++ familyhelper-ui-view/packages/webapp-pc/src/store/modules/navigationEzNav.ts
@@ -354,9 +354,9 @@ function provideStoreSetup(): StoreSetup {
 
 // -----------------------------------------------------------钩子逻辑-----------------------------------------------------------
 // 存储在 Localstorage 中的持久化主键
-const LOCAL_STORAGE_PERSISTENCE_DATA_KEY = 'store.persistence_data.eznav'
+const LOCAL_STORAGE_PERSISTENCE_DATA_KEY = 'store.persistence_data.eznav.v2'
 // 存储在 Settingrepo 中的持久化主键
-const SETTINGREPO_PERSISTENCE_DATA_KEY = 'framework.pc.eznav'
+const SETTINGREPO_PERSISTENCE_DATA_KEY = 'framework.pc.eznav.v2'
 
 let lnpStoreLoginActionHandle: () => void = () => {}
 if (import.meta.hot) {